pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-voice-agent"
version = "1.0.0"
description = "Voice AI agent using Google Gemini 2.5 Flash Native Audio"
requires-python = ">=3.10"

dependencies = [
    "livekit-agents[google]>=1.3.10",
    "livekit-plugins-xai>=1.3.10",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "google-genai",  # For ThinkingConfig
    "httpx",  # For webhook calls
    "boto3",  # For S3 uploads/downloads
    "fastapi",  # For webhook server
    "uvicorn",  # ASGI server
    "openai",  # For Whisper STT (optional)
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
